Move MessageStream back to delegate module
